backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
email-validator==2.2.0
bcrypt==3.2.2
